requirements.txt
# AI Study Companion MVP - Python Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# AWS Services
boto3==1.29.7
botocore==1.32.7

# AI/LLM
openai>=1.6.1,<2.0.0
langchain==0.0.340
langchain-openai==0.0.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2  # For testing FastAPI

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.0
isort==5.12.0

# Logging
structlog==23.2.0

# Data Validation
jsonschema==4.20.0

